Move `useEarnBottomTag` to earn components directory
